make jdl types decoupled
